Start migrating eveything to double

+++ ThreeJs4Net/Math/IVector.cs
@@ -2,32 +2,32 @@
 {
     public interface IVector<T>
     {
-        T SetComponent(int index, float value);
-        public float GetComponent(int index);
+        T SetComponent(int index, double value);
+        public double GetComponent(int index);
         T Copy(T v);
         T Add(T v);
 
         //      set( ...args: number[] ): this;
 
-        T SetScalar(float scalar);
+        T SetScalar(double scalar);
         T AddVectors(T v1, T v2);
-        T AddScaledVector(T v, float s);
-        T AddScalar(float scalar);
+        T AddScaledVector(T v, double s);
+        T AddScalar(double scalar);
         T Sub(T v);
         T SubVectors(T a, T b);
-        T MultiplyScalar(float scalar);
-        T DivideScalar(float scalar);
+        T MultiplyScalar(double scalar);
+        T DivideScalar(double scalar);
         T Negate();
-        float Dot(T v);
-        float LengthSq();
-        float Length();
+        double Dot(T v);
+        double LengthSq();
+        double Length();
         T Normalize();
         //* NOTE: Vector4 doesn't have the property.
-        public float DistanceTo(T vector);
+        public double DistanceTo(T vector);
         //* NOTE: Vector4 doesn't have the property.
-        public float DistanceToSquared(T vector);
-        T SetLength(float length);
-        T Lerp(T v, float alpha);
+        public double DistanceToSquared(T vector);
+        T SetLength(double length);
+        T Lerp(T v, double alpha);
         bool Equals(T v);
         T Clone();
     }
